+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 * Remove `String#escapeUnicode()`
 * Remove `Function#listenTo` and `Function#unlisten`
 * Remove `Object.values()` and `Object.entries()` polyfills
+* Remove `Object.zip()` and `Object.unzip()`
 
 ## 0.7.27 (2022-12-23)